feat: adds ability to enter decimal numbers

See merge request decibel-threshold-event-displayer/decibel-threshold-event-displayer!28

+++ app/js/components/home/home.js
@@ -93,14 +93,14 @@ export class Home extends Component {
     const maxDb = this._select("#maxDb");
     const fileUpload = this._select("#file");
 
-    if (parseInt(minDb.value) >= parseInt(maxDb.value))
+    if (parseFloat(minDb.value) >= parseFloat(maxDb.value))
       maxDb.setCustomValidity("invalid");
     else maxDb.setCustomValidity("");
 
     if (
-      isNaN(parseInt(threshold.value)) ||
-      parseInt(threshold.value) >= parseInt(maxDb.value) ||
-      parseInt(threshold.value) <= 0
+      isNaN(parseFloat(threshold.value)) ||
+      parseFloat(threshold.value) >= parseFloat(maxDb.value) ||
+      parseFloat(threshold.value) <= 0
     )
       threshold.setCustomValidity("invalid");
     else threshold.setCustomValidity("");
